add apis for sensor data

+++ src/model/sensor-data.model.ts
@@ -9,7 +9,7 @@ export class SensorData extends Model {
   temperature: number;
 
   @Column({
-    type: DataType.FLOAT,
+    type: DataType.INTEGER,
     allowNull: false,
   })
   humidity: number;
